Rename Enity Types
