+++ custom_components/hahm/control_unit.py
@@ -44,7 +44,7 @@ from hahomematic.const import (
     HmPlatform,
 )
 from hahomematic.entity import BaseEntity
-from hahomematic.hub import HmHub, HmDummyHub
+from hahomematic.hub import HmDummyHub, HmHub
 from hahomematic.xml_rpc_server import register_xml_rpc_server
 
 from homeassistant.const import CONF_DEVICE_ID
@@ -62,9 +62,8 @@ from .const import (
     ATTR_PATH,
     DOMAIN,
     HAHM_PLATFORMS,
-    HM_ENTITIES,
-    HmCallbackEntity
 )
+from .helpers import HmBaseEntity, HmCallbackEntity
 
 _LOGGER = logging.getLogger(__name__)
 SCAN_INTERVAL = timedelta(seconds=30)
@@ -84,7 +83,7 @@ class ControlUnit:
             control_config.enable_sensors_for_system_variables
         )
         self._central: CentralUnit = self.create_central()
-        self._active_hm_entities: dict[str, HM_ENTITIES] = {}
+        self._active_hm_entities: dict[str, HmBaseEntity] = {}
         self._hub: HaHub | None = None
 
     async def start(self) -> None:
@@ -111,7 +110,9 @@ class ControlUnit:
         await self._central.init_hub()
         self._hub = HaHub(self._hass, self)
         await self._hub.init()
-        hm_entities = [self._central.hub.hub_entities.values()] if self._central.hub else []
+        hm_entities = (
+            [self._central.hub.hub_entities.values()] if self._central.hub else []
+        )
         args = [hm_entities]
 
         async_dispatcher_send(
@@ -171,11 +172,11 @@ class ControlUnit:
 
         return hm_entities
 
-    def add_hm_entity(self, hm_entity: HM_ENTITIES) -> None:
+    def add_hm_entity(self, hm_entity: HmBaseEntity) -> None:
         """add entity to active entities"""
         self._active_hm_entities[hm_entity.unique_id] = hm_entity
 
-    def remove_hm_entity(self, hm_entity: HM_ENTITIES) -> None:
+    def remove_hm_entity(self, hm_entity: HmBaseEntity) -> None:
         """remove entity from active entities"""
         del self._active_hm_entities[hm_entity.unique_id]
 
@@ -328,7 +329,7 @@ class ControlUnit:
             )
         return clients
 
-    def _get_active_entity_by_address(self, address: str) -> HM_ENTITIES | None:
+    def _get_active_entity_by_address(self, address: str) -> HmBaseEntity | None:
         for entity in self._active_hm_entities.values():
             if isinstance(entity, HmCallbackEntity) and entity.address == address:
                 return entity
