+++ custom_components/hahm/entity_helpers.py
@@ -5,7 +5,7 @@ import logging
 from typing import Any
 
 from hahomematic.const import HmPlatform
-from hahomematic.entity import BaseEntity, CustomEntity, GenericEntity
+from hahomematic.entity import CustomEntity, GenericEntity
 
 from homeassistant.components.binary_sensor import (
     BinarySensorDeviceClass,
@@ -39,7 +39,8 @@ from homeassistant.const import (
     VOLUME_CUBIC_METERS,
 )
 from homeassistant.helpers.entity import EntityDescription
-from .generic_entity import HMEntityType
+
+from .helpers import HmGenericEntity
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -435,7 +436,7 @@ _DEFAULT_DESCRIPTION: dict[HmPlatform, Any] = {
 }
 
 
-def get_entity_description(hm_entity: HMEntityType) -> EntityDescription | None:
+def get_entity_description(hm_entity: HmGenericEntity) -> EntityDescription | None:
     """Get the entity_description for platform."""
     entity_description: EntityDescription | None = None
     if isinstance(hm_entity, GenericEntity):
@@ -446,7 +447,7 @@ def get_entity_description(hm_entity: HMEntityType) -> EntityDescription | None:
                 (hm_entity.device_type, hm_entity.parameter)
             )
 
-        if entity_description is None and hm_entity.sub_type :
+        if entity_description is None and hm_entity.sub_type:
             if platform_device_param_descriptions := _ENTITY_DESCRIPTION_DEVICE_PARAM.get(
                 hm_entity.platform
             ):
@@ -474,8 +475,10 @@ def get_entity_description(hm_entity: HMEntityType) -> EntityDescription | None:
             if platform_device_descriptions := _ENTITY_DESCRIPTION_DEVICE.get(
                 hm_entity.platform
             ):
-                entity_description = platform_device_descriptions.get(hm_entity.sub_type)
-                
+                entity_description = platform_device_descriptions.get(
+                    hm_entity.sub_type
+                )
+
         if entity_description:
             return entity_description
